Fixed an issue where the length of plaintext + aead wasn't properly checked

+++ src/main/java/org/unitedid/yhsm/internal/Defines.java
@@ -30,7 +30,7 @@ public class Defines {
     final static public byte YSM_MAX_PKT_SIZE = 0x60;
     final static public int YSM_AEAD_NONCE_SIZE = 6;
     final static public int YSM_AEAD_MAC_SIZE = 8;
-    final static public byte YSM_DATA_BUF_SIZE = 64;
+    final static public int YSM_DATA_BUF_SIZE = 64;
     final static public int YSM_MAX_KEY_SIZE = 32;
     final static public int YSM_AEAD_MAX_SIZE = YSM_DATA_BUF_SIZE + YSM_AEAD_MAC_SIZE;
     final static public int YSM_SHA1_HASH_SIZE = 20;
